+++ src/ISARMIN.Infrastructure/Persistence/Migrations/20260722232907_AgregarImagenProducto.cs
@@ -1,0 +1,28 @@
+﻿using Microsoft.EntityFrameworkCore.Migrations;
+
+#nullable disable
+
+namespace ISARMIN.Infrastructure.Persistence.Migrations
+{
+    /// <inheritdoc />
+    public partial class AgregarImagenProducto : Migration
+    {
+        /// <inheritdoc />
+        protected override void Up(MigrationBuilder migrationBuilder)
+        {
+            migrationBuilder.AddColumn<string>(
+                name: "imagen",
+                table: "productos",
+                type: "text",
+                nullable: true);
+        }
+
+        /// <inheritdoc />
+        protected override void Down(MigrationBuilder migrationBuilder)
+        {
+            migrationBuilder.DropColumn(
+                name: "imagen",
+                table: "productos");
+        }
+    }
+}